WIP more cli stuff
Move "cleanup" into the `cli` dir
