+++ environment-lock.yml
@@ -41,6 +41,7 @@ dependencies:
   - fontconfig=2.13.1=hba837de_1005
   - freetype=2.10.4=h0708190_1
   - freexl=1.0.6=h7f98852_0
+  - funcy=1.16=pyhd8ed1ab_0
   - future=0.18.2=py39hf3d152e_3
   - gdal=3.1.4=py39h3f36f43_3
   - geos=3.8.1=he1b5a44_0
