+++ environment.yml
@@ -20,6 +20,8 @@ dependencies:
   - qgis ~=3.16.0
   - requests ~=2.25.0
   - types-requests ~=2.25.0
+  - pyyaml ~=5.4.1
+  - funcy ~=1.16
 
 
   #############################
